refactor(engine): reduce invocation handler to JVM adapter — Epic 3.7
TramaiInvocationHandler is now a thin JVM proxy adapter (Object methods,
open check, plan resolution, context creation, blocking/suspend lifecycle
bridge, delegation). The per-service invocation execution graph moved to
InvocationExecutionCoordinator (return-kind dispatch + ClaimedResumeExecutor),
with RawResponseCoordinator (STRING/UNIT), ToolLoopCoordinator (provider-tool
loop), and InvocationContextFactory (pre-dispatch context/conversation-ID)
extracted. The suspend lifecycle bridge (synchronized active-invocation
registry, exactly-once resume) is preserved verbatim; zero intended behavior
change — 20/20 characterization traces byte-identical.
