+++ tramai-engine/src/main/kotlin/dev/tramai/engine/TramaiEngine.kt
@@ -158,6 +158,8 @@ import dev.tramai.engine.planning.OperationDefinitionCompiler
 import dev.tramai.engine.planning.OperationFingerprintFactory
 import dev.tramai.engine.planning.ServiceDefinition
 import dev.tramai.engine.planning.ServiceDefinitionCompiler
+import dev.tramai.engine.invocation.InvocationExecutionCoordinator
+import dev.tramai.engine.invocation.TramaiInvocationHandler
 import dev.tramai.core.provider.resolveCandidates
 
 private const val UNREGISTERED_TOOL_NAME = "unregistered_tool"
@@ -537,18 +539,25 @@ class TramaiEngine private constructor(
     fun <T : Any> create(serviceType: KClass<T>): T {
         check(!closed.get()) { "Tramai runtime is closed" }
         val definition = serviceDefinitionCompiler.compile(serviceType)
-        val handler = TramaiInvocationHandler(
+        val executionCoordinator = InvocationExecutionCoordinator(
             components = components,
             circuitBreaker = circuitBreaker,
             retryDelayPolicy = retryDelayPolicy,
             migrationWarningGuard = migrationWarningGuard,
+            lifecycleScope = lifecycleScope,
+            isClosed = closed,
+            serviceDefinition = definition,
+            resumeOperationRegistry = resumeOperationRegistry,
+        )
+        val handler = TramaiInvocationHandler(
+            serviceDefinition = definition,
             lifecycleJob = lifecycleJob,
             lifecycleScope = lifecycleScope,
             isClosed = closed,
             engineThreadMarker = engineThreadMarker,
             activeInvocationJobs = activeInvocationJobs,
-            serviceDefinition = definition,
-            resumeOperationRegistry = resumeOperationRegistry,
+            contextFactory = executionCoordinator.contextFactory,
+            executionCoordinator = executionCoordinator,
         )
 
         @Suppress("UNCHECKED_CAST")
@@ -559,7 +568,7 @@ class TramaiEngine private constructor(
         ) as T).also {
             resumeOperationRegistry.registerAll(
                 serviceDefinition = definition,
-                resumeExecutor = handler,
+                resumeExecutor = executionCoordinator,
             )
         }
     }
@@ -575,22 +584,19 @@ class TramaiEngine private constructor(
     fun registerService(serviceType: KClass<*>) {
         check(!closed.get()) { "Tramai runtime is closed" }
         val definition = serviceDefinitionCompiler.compile(serviceType)
-        val handler = TramaiInvocationHandler(
+        val executionCoordinator = InvocationExecutionCoordinator(
             components = components,
             circuitBreaker = circuitBreaker,
             retryDelayPolicy = retryDelayPolicy,
             migrationWarningGuard = migrationWarningGuard,
-            lifecycleJob = lifecycleJob,
             lifecycleScope = lifecycleScope,
             isClosed = closed,
-            engineThreadMarker = engineThreadMarker,
-            activeInvocationJobs = activeInvocationJobs,
             serviceDefinition = definition,
             resumeOperationRegistry = resumeOperationRegistry,
         )
         resumeOperationRegistry.registerAll(
             serviceDefinition = definition,
-            resumeExecutor = handler,
+            resumeExecutor = executionCoordinator,
         )
     }
 
@@ -679,1042 +685,6 @@ data class ResumeApprovalCommand(
     val presentedToken: ApprovalToken,
     val resumedBy: String,
 )
-
-internal class TramaiInvocationHandler(
-    private val components: EngineComponents,
-    private val circuitBreaker: ProviderCircuitBreaker,
-    private val retryDelayPolicy: ProviderRetryDelayPolicy,
-    private val migrationWarningGuard: java.util.concurrent.atomic.AtomicBoolean,
-    private val lifecycleJob: Job,
-    private val lifecycleScope: CoroutineScope,
-    private val isClosed: java.util.concurrent.atomic.AtomicBoolean = java.util.concurrent.atomic.AtomicBoolean(false),
-    private val engineThreadMarker: ThreadLocal<Boolean> = ThreadLocal(),
-    private val activeInvocationJobs: MutableSet<Job> = java.util.concurrent.ConcurrentHashMap.newKeySet(),
-    private val serviceDefinition: ServiceDefinition,
-    private val resumeOperationRegistry: ResumeOperationRegistry,
-) : InvocationHandler, ClaimedResumeExecutor {
-
-    private val routingPlan = components.providers.routingPlan
-    private val structuredOutputHandler = components.execution.structuredOutputHandler
-    private val toolRegistry = components.tools.toolRegistry
-    private val operationObserver = components.observation.operationObserver
-    private val operationInterceptor = components.observation.operationInterceptor
-    private val responseCache = components.persistence.responseCache
-    private val modelRegistry = components.security.modelRegistry
-    private val modelRegistrySettings = components.security.modelRegistrySettings
-    private val tokenBudgetSettings = components.execution.tokenBudgetSettings
-    private val promptSanitizer = components.security.promptSanitizer
-    private val chatMemory = components.persistence.chatMemory
-    private val conversationIdProvider = components.persistence.conversationIdProvider
-    private val dlpInterceptor = components.security.dlpInterceptor
-    private val dlpRedactionAuditEmitter = components.security.dlpRedactionAuditEmitter
-    private val toolResultFilteringSettings = components.tools.toolResultFilteringSettings
-    private val engineEventObserver = components.observation.engineEventObserver
-    private val toolFailureDiagnosticObserver = components.observation.toolFailureDiagnosticObserver
-    private val structuredOutputFailureDiagnosticObserver = components.observation.structuredOutputFailureDiagnosticObserver
-    private val policyDecisionAuditEmitter = components.security.policyDecisionAuditEmitter
-    private val suspendedInvocationStore = components.approvals.suspendedInvocationStore
-    private val approvalLifecycleAuditEmitter = components.approvals.approvalLifecycleAuditEmitter
-    private val clock = components.execution.clock
-    private val approvalContinuationStore = (components.approvals.capability as? ApprovalCapability.Enabled)?.continuationStore
-    private val toolArgumentsDigester = (components.approvals.capability as? ApprovalCapability.Enabled)?.argumentsDigester
-    private val approvalGateCoordinator = (components.approvals.capability as? ApprovalCapability.Enabled)?.gateCoordinator
-
-    private val policyHelper = PolicyEnforcementHelper(components.security.resolvedPolicyEngine, migrationWarningGuard, isLegacyFallback = components.security.isLegacyFallback, auditEmitter = policyDecisionAuditEmitter)
-    private val modelRegistryEnforcer = ModelRegistryEnforcer(modelRegistry, modelRegistrySettings)
-    private val beforeProviderInvocationGate = ProviderInvocationGate { providerId, modelName, correlationId, securityContext -> enforceBeforeProviderInvocation(providerId, modelName, correlationId, securityContext) }
-    private val beforeResolutionGate = ProviderResolutionGate { operation, correlationId, securityContext -> enforceBeforeProviderResolution(operation, correlationId, securityContext) }
-    private val fallbackGate = ProviderFallbackGate { correlationId, previousProviderId, previousModelName, nextProviderId, reason, securityContext -> enforceFallbackTransition(correlationId, previousProviderId, previousModelName, nextProviderId, reason, securityContext) }
-    private val providerExecutionCoordinator = ProviderExecutionCoordinator(
-        routingPlan = routingPlan,
-        circuitBreaker = circuitBreaker,
-        attemptExecutor = ProviderAttemptExecutor(
-            serviceInterface = serviceDefinition.serviceType.qualifiedName ?: serviceDefinition.serviceType.simpleName.orEmpty(),
-            operationObserver = operationObserver,
-            operationInterceptor = operationInterceptor,
-            circuitBreaker = circuitBreaker,
-            retryPolicy = ProviderRetryPolicy(retryDelayPolicy),
-            authorizationService = ProviderAuthorizationService(modelRegistryEnforcer),
-            beforeProviderInvocation = beforeProviderInvocationGate,
-            responseSanitizer = ProviderResponseSanitizer { response, operation, providerId, modelName, correlationId, securityContext, observation -> sanitizeProviderResponse(response, operation, providerId, modelName, correlationId, securityContext, observation) },
-        ),
-        fallbackPolicy = ProviderFallbackPolicy(),
-        beforeResolution = beforeResolutionGate,
-        fallbackGate = fallbackGate,
-    )
-    private val toolExposureCoordinator = ToolExposureCoordinator(toolRegistry, policyHelper)
-    private val conversationMemoryCoordinator = ConversationMemoryCoordinator(
-        chatMemory = chatMemory,
-        conversationIdProvider = conversationIdProvider,
-    )
-    private val operationCacheCoordinator = OperationCacheCoordinator(
-        responseCache = responseCache,
-        operationInterceptor = operationInterceptor,
-        dlpInterceptor = dlpInterceptor,
-        modelRegistrySettings = modelRegistrySettings,
-        modelRegistryEnforcer = modelRegistryEnforcer,
-        policyHelper = policyHelper,
-    )
-    private val tokenBudgetCoordinator = TokenBudgetCoordinator(tokenBudgetSettings)
-    private val streamingExecutionCoordinator = StreamingExecutionCoordinator(
-        routingPlan = routingPlan,
-        circuitBreaker = circuitBreaker,
-        lifecycleScope = lifecycleScope,
-        isClosed = isClosed,
-        serviceTypeName = serviceDefinition.serviceType.qualifiedName ?: serviceDefinition.serviceType.simpleName.orEmpty(),
-        qualifiedServiceName = serviceDefinition.serviceType.qualifiedName,
-        operationObserver = operationObserver,
-        operationInterceptor = operationInterceptor,
-        toolExposureCoordinator = toolExposureCoordinator,
-        conversationMemoryCoordinator = conversationMemoryCoordinator,
-        tokenBudgetCoordinator = tokenBudgetCoordinator,
-        modelRegistryEnforcer = modelRegistryEnforcer,
-        beforeResolution = beforeResolutionGate,
-        beforeInvocation = beforeProviderInvocationGate,
-        fallbackGate = fallbackGate,
-        beforeResponseReturn = StreamingBeforeResponseReturnGate { route, correlationId, securityContext ->
-            enforceBeforeResponseReturn(route, correlationId, securityContext)
-        },
-    )
-    private val toolResultSanitizer = ToolResultSanitizer(
-        toolRegistry = toolRegistry,
-        dlpInterceptor = dlpInterceptor,
-        dlpRedactionAuditEmitter = dlpRedactionAuditEmitter,
-        toolResultFilteringSettings = toolResultFilteringSettings,
-        engineEventObserver = engineEventObserver,
-    )
-    private val toolAuthorizationCoordinator = ToolAuthorizationCoordinator(policyHelper)
-    private val toolRetryPolicy = ToolRetryPolicy()
-    private val approvalSuspensionCoordinator = ApprovalSuspensionCoordinator(
-        approvalGateCoordinator = approvalGateCoordinator,
-        approvalContinuationStore = approvalContinuationStore,
-        suspendedInvocationStore = suspendedInvocationStore,
-        resumeOperationRegistry = resumeOperationRegistry,
-        serviceDefinition = serviceDefinition,
-        resumeExecutor = this,
-        toolArgumentsDigester = toolArgumentsDigester,
-        clock = clock,
-        approvalLifecycleAuditEmitter = approvalLifecycleAuditEmitter,
-    )
-    private val toolInvocationExecutor = ToolInvocationExecutor(
-        authorizationCoordinator = toolAuthorizationCoordinator,
-        retryPolicy = toolRetryPolicy,
-        toolFailureDiagnosticObserver = toolFailureDiagnosticObserver,
-        approvalGate = approvalSuspensionCoordinator,
-    )
-    private val toolReinjectionCoordinator = ToolReinjectionCoordinator(
-        toolRegistry = toolRegistry,
-        policyHelper = policyHelper,
-        invocationExecutor = toolInvocationExecutor,
-        resultSanitizer = toolResultSanitizer,
-    )
-    private val structuredResponseCoordinator = StructuredResponseCoordinator(
-        structuredOutputHandler = structuredOutputHandler,
-        structuredOutputFailureDiagnosticObserver = structuredOutputFailureDiagnosticObserver,
-        conversationMemoryCoordinator = conversationMemoryCoordinator,
-        operationCacheCoordinator = operationCacheCoordinator,
-        policyHelper = policyHelper,
-        attemptExecutor = StructuredAttemptExecutor { request ->
-            executeWithTools(
-                ToolLoopContext(
-                    operation = request.operation,
-                    messages = request.messages,
-                    tokenBudgetTracker = request.tokenBudgetTracker,
-                    correlationId = request.correlationId,
-                    securityContext = request.securityContext,
-                    identity = request.identity,
-                    conversationId = request.conversationId,
-                    historySize = request.historySize,
-                ),
-            )
-        },
-        serviceTypeName = serviceDefinition.serviceType.qualifiedName
-            ?: serviceDefinition.serviceType.simpleName
-            ?: "<unknown>",
-    )
-
-    private fun OperationObservation.completeCancellation(cancellation: CancellationException) {
-        try {
-            onCallCancelled()
-        } catch (observerError: Throwable) {
-            cancellation.addSuppressed(observerError)
-        }
-    }
-
-    private suspend fun enforceBeforeProviderResolution(
-        operation: OperationDefinition,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-    ) {
-        policyHelper.enforce(
-            policyHelper.buildContext(
-                enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_PROVIDER_RESOLUTION,
-                correlationId = correlationId,
-            ).modelName(operation.operation.model)
-                .applySecurityContext(securityContext)
-                .build()
-        )
-    }
-
-    private suspend fun enforceBeforeResponseReturn(
-        route: ResolvedProviderRoute,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-    ) {
-        policyHelper.enforce(
-            policyHelper.buildContext(
-                enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_RESPONSE_RETURN,
-                correlationId = correlationId,
-            ).providerId(route.providerName)
-                .modelName(route.effectiveModelName)
-                .applySecurityContext(securityContext)
-                .build()
-        )
-    }
-
-    private suspend fun enforceBeforeProviderInvocation(
-        providerId: String,
-        modelName: String,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-    ) {
-        policyHelper.enforce(
-            policyHelper.buildContext(
-                enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_PROVIDER_INVOCATION,
-                correlationId = correlationId,
-            ).providerId(providerId)
-                .modelName(modelName)
-                .applySecurityContext(securityContext)
-                .build()
-        )
-    }
-
-    /**
-     * Enforces [EnforcementPoint.BEFORE_FALLBACK] at any transition point.
-     * Used for provider failures, circuit breaker opens, streaming startup failures,
-     * and route-unavailable transitions.
-     */
-    private suspend fun enforceFallbackTransition(
-        correlationId: String,
-        previousProviderId: String?,
-        previousModelName: String?,
-        nextProviderId: String,
-        reason: String,
-        securityContext: ExecutionSecurityContext,
-    ) {
-        val ctx = policyHelper.buildContext(
-            enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_FALLBACK,
-            correlationId = correlationId,
-        ).applySecurityContext(securityContext)
-        if (previousProviderId != null) ctx.providerId(previousProviderId)
-        if (previousModelName != null) ctx.modelName(previousModelName)
-        ctx.fallbackProviderId(nextProviderId)
-        ctx.attribute("fallbackReason", reason)
-        policyHelper.enforce(ctx.build())
-    }
-
-    /**
-     * Finalises tool-processing observation without a suspend boundary,
-     * so the cancellation scanner does not flag a broad [Throwable] catch.
-     *
-     * When [primaryError] is provided, observer failure is suppressed on it.
-     * When it is null (successful tool processing), observer failure is
-     * logged as a warning but does not invalidate the completed side effect.
-     */
-    private fun OperationObservation.completeAfterToolProcessing(
-        primaryError: Throwable? = null,
-    ) {
-        try {
-            onCallCompleted(parseSuccess = null)
-        } catch (observerError: Throwable) {
-            if (primaryError != null) {
-                primaryError.addSuppressed(observerError)
-            } else {
-                System.getLogger("dev.tramai.engine.TramaiEngine").log(
-                    System.Logger.Level.WARNING,
-                    "Operation observer failed after successful tool processing",
-                    observerError,
-                )
-            }
-        }
-    }
-
-    override fun invoke(proxy: Any, method: Method, args: Array<out Any?>?): Any? {
-        if (method.declaringClass == Any::class.java) {
-            return handleObjectMethod(proxy, method, args.orEmpty())
-        }
-
-        check(!isClosed.get()) { "Tramai runtime is closed" }
-
-        val plan = serviceDefinition.operations[method]
-            ?: throw ConfigurationException("No operation metadata registered for ${method.name}")
-        val operation = plan.definition
-
-        val conversationId = if (chatMemory != null) {
-            conversationMemoryCoordinator.resolveConversationId(method, args.orEmpty())
-        } else {
-            null
-        }
-        return if (operation.isSuspend) {
-            invokeSuspend(operation, args.orEmpty(), conversationId)
-        } else {
-            // Run the blocking call as a child of the engine's OWN lifecycle
-            // job (not the caller-supplied job/scope): close() cancels and
-            // joins lifecycleJob, so it can terminate a blocking provider
-            // that is still executing when the engine is closed. The thread
-            // marker marks this coroutine as engine-owned so a blocking call
-            // that itself invokes close() skips the join (avoiding a
-            // self-deadlock on lifecycleJob).
-            val result = runBlocking(lifecycleJob + engineThreadMarker.asContextElement(true)) {
-                execute(operation, args.orEmpty().toList(), conversationId)
-            }
-            // The engine may have closed while this blocking call was in
-            // flight. Never deliver a result computed against a closed engine:
-            // the caller sees the fixed lifecycle error instead.
-            check(!isClosed.get()) { "Tramai runtime is closed" }
-            result
-        }
-    }
-
-    private fun invokeSuspend(
-        operation: OperationDefinition,
-        args: Array<out Any?>,
-        conversationId: String?,
-    ): Any {
-        // Kotlin suspend proxies receive the continuation as the last JVM argument.
-        @Suppress("UNCHECKED_CAST")
-        val continuation = args.lastOrNull() as? Continuation<Any?>
-            ?: throw ConfigurationException("Suspend invocation for ${operation.method.name} is missing its continuation")
-
-        val callArguments = args.dropLast(1)
-        // Launch as a child of the CALLER's job (continuation.context, with the
-        // caller's Job element retained) so parent cancellation propagates
-        // synchronously into the in-flight invocation (validated by the
-        // ToolSafeFailureContract / StructuredOutputFailureBoundary
-        // parent-cancellation tests). The invocation RUNS on the engine's own
-        // dispatcher (lifecycleScope), NOT the caller's: if it ran on the
-        // caller's single-threaded dispatcher, close() joining it could
-        // deadlock when that thread is blocked inside close(). Engine close()
-        // owns the work: the launch+add is synchronized with close()'s cancel
-        // snapshot, the closed flag is re-checked INSIDE the lock, and close()
-        // cancels AND joins every tracked invocation. Exactly-once resume: the
-        // block records the outcome BEFORE resuming the continuation, and
-        // invokeOnCompletion resumes with a cancellation when the block never
-        // ran (job cancelled pre-start by close()) — otherwise the caller's
-        // suspension would freeze forever.
-        val resumed = java.util.concurrent.atomic.AtomicReference<Result<Any?>?>(null)
-        val launched = synchronized(activeInvocationJobs) {
-            check(!isClosed.get()) { "Tramai runtime is closed" }
-            val job = lifecycleScope.launch(
-                continuation.context.minusKey(kotlin.coroutines.ContinuationInterceptor) +
-                    engineThreadMarker.asContextElement(true),
-            ) {
-                var result = runCatching { execute(operation, callArguments, conversationId) }
-                // Never deliver a success computed against a closed engine: the
-                // engine may have closed while the invocation was in flight.
-                // The caller sees the fixed lifecycle error instead (mirrors
-                // the blocking path).
-                if (isClosed.get() && result.isSuccess) {
-                    result = Result.failure(IllegalStateException("Tramai runtime is closed"))
-                }
-                resumed.set(result)
-                continuation.resumeWith(result)
-            }
-            activeInvocationJobs += job
-            job
-        }
-        launched.invokeOnCompletion { cause ->
-            // Registry mutations obey the same monitor: launch+add (above) and
-            // close()'s snapshot (in close()) are synchronized on
-            // activeInvocationJobs, so removal must be too. An unsynchronized
-            // removal can race Kotlin's toList() size-1 fast path inside the
-            // snapshot (size()==1, then remove, then iterator().next() throws
-            // NoSuchElementException), making close() throw and the closer
-            // thread die silently — the original CI hang.
-            synchronized(activeInvocationJobs) {
-                activeInvocationJobs -= launched
-            }
-            if (resumed.get() == null) {
-                // Block never ran (e.g. cancelled before the dispatcher started
-                // it): resume so the caller's suspension does not freeze.
-                continuation.resumeWith(
-                    Result.failure(
-                        cause as? CancellationException ?: CancellationException("Engine closed", cause),
-                    ),
-                )
-            }
-        }
-        return COROUTINE_SUSPENDED
-    }
-
-    private suspend fun execute(
-        operation: OperationDefinition,
-        arguments: List<Any?>,
-        conversationId: String?,
-    ): Any? {
-        val tokenBudgetTracker = tokenBudgetCoordinator.createTracker()
-        val workflowRunId = java.util.UUID.randomUUID().toString()
-        val workflowDigest = WorkflowDigestHelper.compute(operation, serviceDefinition)
-        val identity = EngineExecutionIdentity(
-            workflowRunId = workflowRunId,
-            correlationId = "", // Will be set in executeRaw/the structured coordinator
-            workflowDigest = workflowDigest,
-            policyVersion = policyHelper.getPolicyVersion(),
-            actorId = PolicyEnforcementHelper.ACTOR_ANONYMOUS,
-        )
-        return when (operation.returnKind) {
-            ReturnKind.STRING -> executeRaw(operation, arguments, tokenBudgetTracker, conversationId, identity)
-            ReturnKind.UNIT -> {
-                executeRaw(operation, arguments, tokenBudgetTracker, conversationId, identity)
-                Unit
-            }
-            ReturnKind.STRUCTURED -> structuredResponseCoordinator.execute(
-                StructuredResponseRequest(
-                    operation = operation,
-                    arguments = arguments,
-                    tokenBudgetTracker = tokenBudgetTracker,
-                    conversationId = conversationId,
-                    identity = identity,
-                    operationFingerprint = serviceDefinition.operations[operation.method]?.fingerprint,
-                ),
-            )
-            ReturnKind.STREAMING -> streamingExecutionCoordinator.execute(
-                StreamingExecutionRequest(
-                    operation = operation,
-                    arguments = arguments,
-                    tokenBudgetTracker = tokenBudgetTracker,
-                    conversationId = conversationId,
-                ),
-            )
-        }
-    }
-
-    private suspend fun executeRaw(
-        operation: OperationDefinition,
-        arguments: List<Any?>,
-        tokenBudgetTracker: TokenBudgetTracker,
-        conversationId: String?,
-        identity: EngineExecutionIdentity,
-    ): String {
-        val securityContext = ExecutionSecurityContext.fromArguments(arguments.toTypedArray())
-        val correlationId = java.util.UUID.randomUUID().toString()
-        val effectiveIdentity = identity.copy(correlationId = correlationId)
-        val initialMessages = operation.initialMessages(arguments)
-        val prepared = conversationMemoryCoordinator.prepareMessages(initialMessages, conversationId)
-        val history = prepared?.history ?: emptyList()
-        val effectiveMessages = prepared?.effectiveMessages ?: initialMessages
-        val cacheKey = operationCacheCoordinator.createKey(
-            OperationCacheKeyRequest(
-                digestSource = effectiveMessages,
-                securityPartition = securityContext.toCacheSecurityPartition(),
-                operationFingerprint = serviceDefinition.operations[operation.method]?.fingerprint,
-                requestedModel = operation.operation.model,
-                explicitProvider = operation.operation.provider.takeIf { it.isNotBlank() },
-                serviceInterface = operation.method.declaringClass.name,
-                methodName = operation.method.name,
-                toolDefinitions = operation.toolDefinitions,
-                operation = operation.operation,
-                returnKind = operation.returnKind,
-                conversationId = conversationId,
-            ),
-        )
-        if (cacheKey != null) {
-            when (val cached = operationCacheCoordinator.lookup(
-                OperationCacheLookupRequest(cacheKey, securityContext, correlationId, conversationId),
-            )) {
-                is OperationCacheLookupResult.Hit -> return cached.value as String
-                is OperationCacheLookupResult.Miss -> Unit
-            }
-        }
-
-        val effectiveMutableMessages = effectiveMessages.toMutableList()
-
-        val result = executeWithTools(
-            ToolLoopContext(
-                operation = operation,
-                messages = effectiveMutableMessages,
-                tokenBudgetTracker = tokenBudgetTracker,
-                correlationId = correlationId,
-                securityContext = securityContext,
-                identity = effectiveIdentity,
-                conversationId = conversationId,
-                historySize = history.size,
-            ),
-        )
-
-        // DLP is already applied inside ProviderAttemptExecutor — use the sanitized response directly
-
-        // Enforce BEFORE_RESPONSE_RETURN
-        policyHelper.enforce(
-            policyHelper.buildContext(
-                enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_RESPONSE_RETURN,
-                correlationId = correlationId,
-            ).providerId(result.providerId)
-                .modelName(result.modelName)
-                .applySecurityContext(securityContext)
-                .build()
-        )
-
-        // Memory: persist response if chatMemory is configured
-        if (conversationId != null) {
-            val assistantMessage = Message(
-                role = MessageRole.ASSISTANT,
-                content = result.response.content,
-                toolCalls = result.response.toolCalls,
-            )
-            conversationMemoryCoordinator.persistTurn(
-                PersistConversationTurnRequest(conversationId, effectiveMutableMessages, history.size, assistantMessage),
-            )
-        }
-
-        result.observation.onCallCompleted(parseSuccess = null)
-        return result.response.content.also {
-            cacheKey?.let { key ->
-                operationCacheCoordinator.store(
-                    OperationCacheStoreRequest(key, it, result.providerId, result.modelName, securityContext, conversationId, result.approvedModel, operation.operation.cacheTtlMillis),
-                )
-            }
-        }
-    }
-
-    private suspend fun executeWithTools(
-        context: ToolLoopContext,
-    ): ProviderCallResult {
-        val operation = context.operation
-        val messages = context.messages
-        val tokenBudgetTracker = context.tokenBudgetTracker
-        val correlationId = context.correlationId
-        val securityContext = context.securityContext
-        val maxToolLoops = 5 // Guard against infinite tool loops
-        val attemptCounter = AttemptCounter()
-        repeat(maxToolLoops) {
-            val result = providerExecutionCoordinator.execute(
-                ProviderExecutionRequest(
-                    operation = operation,
-                    messages = messages,
-                    attemptCounter = attemptCounter,
-                    correlationId = correlationId,
-                    securityContext = securityContext,
-                    beforeRoute = { toolExposureCoordinator.enforce(operation, correlationId, securityContext) },
-                ),
-            )
-            try {
-                tokenBudgetCoordinator.enforce(
-                    tracker = tokenBudgetTracker,
-                    response = result.response,
-                    observation = result.observation,
-                    providerId = result.providerId,
-                    modelName = result.modelName,
-                )
-            } catch (error: TokenBudgetExceededException) {
-                result.observation.onCallCompleted(parseSuccess = null)
-                throw error
-            }
-
-            val toolCalls = result.response.toolCalls
-            if (toolCalls.isNullOrEmpty()) {
-                return result
-            }
-
-            // Normalize unregistered tool calls: replace unknown names with safe placeholder
-            val normalizedToolCalls = toolCalls.map { toolCall ->
-                if (toolRegistry.resolve(toolCall.name) == null) {
-                    toolCall.copy(name = UNREGISTERED_TOOL_NAME, argumentsJson = "{}")
-                } else {
-                    toolCall
-                }
-            }
-
-            // Append assistant message with normalized tool calls
-            messages += Message(
-                role = MessageRole.ASSISTANT,
-                content = result.response.content,
-                toolCalls = normalizedToolCalls,
-            )
-
-            // Tool execution must complete before the observation is finalised,
-            // so that cancellation during tool execution calls onCallCancelled
-            // instead of onCallCompleted.
-            // The try covers only tool reinjection, not onCallCompleted:
-            // if the observer throws after successful tool execution, that
-            // failure is suppressed on the process error rather than causing
-            // a duplicate onCallCompleted call or invalidating the side effect.
-            try {
-                toolReinjectionCoordinator.process(
-                    ToolCallBatchRequest(
-                        operation = operation,
-                        messages = messages,
-                        toolCalls = normalizedToolCalls,
-                        correlationId = correlationId,
-                        securityContext = securityContext,
-                        identity = context.identity,
-                        tokenBudgetTracker = tokenBudgetTracker,
-                        conversationId = context.conversationId,
-                        historySize = context.historySize,
-                        resumingApproval = context.resumingApproval,
-                        parentApprovalId = context.parentApprovalId,
-                    ),
-                )
-            } catch (cancellation: CancellationException) {
-                result.observation.completeCancellation(cancellation)
-                throw cancellation
-            } catch (error: Throwable) {
-                error.rethrowIfCancellation()
-
-                // Suppress observer failure on the process error so that
-                // a failing onCallCompleted cannot duplicate the callback,
-                // and this non-suspend helper keeps the cancellation scanner
-                // satisfied.
-                result.observation.completeAfterToolProcessing(primaryError = error)
-
-                throw error
-            }
-
-            result.observation.completeAfterToolProcessing()
-        }
-        error("Exceeded maximum tool call loops ($maxToolLoops)")
-    }
-
-    private data class ToolLoopContext(
-        val operation: OperationDefinition,
-        val messages: MutableList<Message>,
-        val tokenBudgetTracker: TokenBudgetTracker,
-        val correlationId: String,
-        val securityContext: ExecutionSecurityContext,
-        val identity: EngineExecutionIdentity,
-        val conversationId: String? = null,
-        val historySize: Int = 0,
-        val resumingApproval: Boolean = false,
-        val parentApprovalId: String? = null,
-    )
-
-    /**
-     * Performs the authoritative DLP scan for a single response boundary.
-     *
-     * `text` must be the raw pre-DLP text for the specific scan boundary being inspected.
-     */
-    private suspend fun inspectDlpAuthoritatively(
-        context: DlpContext,
-        text: String,
-    ) = dlpInterceptor.inspect(context, text).also { result ->
-        val sanitizedTextChanged = result.sanitizedText != text
-        val hasRedactionEvidence = result.redactions.isNotEmpty()
-
-        if (sanitizedTextChanged && !hasRedactionEvidence && dlpRedactionAuditEmitter !== NoOpDlpRedactionAuditEmitter) {
-            throw DlpInspectionException("DLP modified output without redaction evidence")
-        }
-        if (!sanitizedTextChanged && hasRedactionEvidence) {
-            throw DlpInspectionException("DLP redactions reported without modifying output")
-        }
-        if (result.redactions.isNotEmpty()) {
-            try {
-                dlpRedactionAuditEmitter.emit(context, result.redactions)
-            } catch (error: CancellationException) {
-                throw error
-            } catch (error: Exception) {
-                error.rethrowIfCancellation()
-                throw DlpInspectionException(
-                    message = "DLP redaction audit emission failed",
-                    cause = error,
-                )
-            }
-        }
-    }
-
-    private fun inspectDlpForDetectionOnly(
-        context: DlpContext,
-        text: String,
-    ) = dlpInterceptor.inspect(context, text)
-
-    /**
-     * Applies authoritative DLP inspection to model output without marking failures as provider failures.
-     */
-    private suspend fun sanitizeProviderResponse(
-        interceptedResponse: ModelResponse,
-        operation: OperationDefinition,
-        providerId: String,
-        modelName: String,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-        observation: OperationObservation,
-    ): ModelResponse = try {
-        if (dlpInterceptor === NoOpDlpInterceptor) {
-            interceptedResponse
-        } else {
-            applyProviderOutputDlp(
-                interceptedResponse = interceptedResponse,
-                operation = operation,
-                providerId = providerId,
-                modelName = modelName,
-                correlationId = correlationId,
-                securityContext = securityContext,
-            )
-        }
-    } catch (e: CancellationException) {
-        throw e
-    } catch (e: DlpInspectionException) {
-        throw e
-    } catch (e: Exception) {
-        e.rethrowIfCancellation()
-        observation.onEngineEvent(
-            name = "tramai.dlp.inspection_failed",
-            attributes = mapOf("providerId" to providerId, "correlationId" to correlationId),
-        )
-        throw DlpInspectionException(
-            message = "DLP inspection failed for provider '$providerId'",
-            cause = e,
-        )
-    }
-
-    /**
-     * Builds the model-output DLP context and returns the sanitized response.
-     */
-    private suspend fun applyProviderOutputDlp(
-        interceptedResponse: ModelResponse,
-        operation: OperationDefinition,
-        providerId: String,
-        modelName: String,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-    ): ModelResponse {
-        val dlpContext = DlpContext(
-            contentType = DlpContentType.MODEL_OUTPUT,
-            contentLocation = DlpContentLocation.MODEL_RESPONSE_CONTENT,
-            operationInterface = serviceDefinition.serviceType.qualifiedName
-                ?: serviceDefinition.serviceType.simpleName.orEmpty(),
-            operationMethod = operation.method.name,
-            providerId = providerId,
-            modelName = modelName,
-            correlationId = correlationId,
-            dataClassification = securityContext.dataClassification,
-            classificationSource = securityContext.classificationSource,
-        )
-        val dlpResult = inspectDlpAuthoritatively(dlpContext, interceptedResponse.content)
-        return if (dlpResult.sanitizedText != interceptedResponse.content) {
-            interceptedResponse.copy(content = dlpResult.sanitizedText)
-        } else {
-            interceptedResponse
-        }
-    }
-
-    override suspend fun execute(request: ClaimedResumeExecutionRequest): Any? {
-        val metadata = request.metadata
-        val registered = request.registered
-        val tokenBudgetTracker = tokenBudgetCoordinator.restoreTracker(metadata.tokenBudgetSnapshot)
-        val toolResult = executeResumedTool(
-            request = request,
-            tokenBudgetTracker = tokenBudgetTracker,
-        )
-        val messages = request.rehydratedPayload.messages.toMutableList()
-        val loopResult = continueAfterToolResult(
-            ContinueAfterToolResultRequest(
-                operation = registered.operation,
-                messages = messages,
-                toolResult = toolResult,
-                toolCallId = metadata.toolCallId,
-                toolCallIndex = metadata.toolCallIndex,
-                correlationId = metadata.correlationId,
-                securityContext = metadata.securityContext,
-                identity = metadata.identity,
-                tokenBudgetTracker = tokenBudgetTracker,
-                suspendedToolName = metadata.toolName,
-                approvalId = request.command.approvalId,
-                conversationId = metadata.conversationId,
-                historySize = metadata.historySize,
-                resumingApproval = true,
-            ),
-        )
-        return finalizeResumedOperation(
-            operation = registered.operation,
-            loopResult = loopResult,
-            messages = messages,
-            correlationId = metadata.correlationId,
-            securityContext = metadata.securityContext,
-            conversationId = metadata.conversationId,
-            historySize = metadata.historySize,
-        )
-    }
-
-    private suspend fun executeResumedTool(
-        request: ClaimedResumeExecutionRequest,
-        tokenBudgetTracker: TokenBudgetTracker,
-    ): ToolResult {
-        val command = request.command
-        val metadata = request.metadata
-        val registered = request.registered
-        val resolvedTool = request.resolvedTool
-        val rehydratedPayload = request.rehydratedPayload
-        val validatedInput = request.validatedInput
-        val expectedArgsDigest = request.expectedArgsDigest
-        val validatedToolCall = dev.tramai.core.model.ToolCall(
-            id = metadata.toolCallId,
-            name = metadata.toolName,
-            argumentsJson = validatedInput,
-        )
-        approvalLifecycleAuditEmitter.onToolExecutionResumed(
-            approvalId = command.approvalId,
-            workflowRunId = metadata.identity.workflowRunId,
-            toolName = metadata.toolName,
-            resumedBy = command.resumedBy,
-        )
-        return try {
-            toolInvocationExecutor.execute(
-                ToolExecutionRequest(
-                    tool = resolvedTool,
-                    toolCall = validatedToolCall,
-                    operation = registered.operation,
-                    correlationId = metadata.correlationId,
-                    securityContext = metadata.securityContext,
-                    identity = metadata.identity,
-                    messages = rehydratedPayload.messages,
-                    tokenBudgetTracker = tokenBudgetTracker,
-                    conversationId = metadata.conversationId,
-                    historySize = metadata.historySize,
-                    resumingApproval = true,
-                    parentApprovalId = command.approvalId,
-                    idempotencyKey = IdempotencyKeyUtil.deriveApprovalKey(command.approvalId, metadata.toolCallId, expectedArgsDigest),
-                    allowRenewedApprovedBindingDuringResume = true,
-                ),
-            )
-        } catch (e: dev.tramai.core.exception.NestedApprovalNotSupportedException) {
-            throw e
-        } catch (e: ToolInvalidInputException) {
-            request.uncertainOutcomeEmitter("tool-execution-failed: ${e::class.simpleName ?: "unknown"}")
-            throw e
-        }
-    }
-
-    /**
-     * Finalizes a resumed operation for all return kinds that don't need
-     * structured parsing. Enforces BEFORE_RESPONSE_RETURN, persists conversation
-     * memory, completes the observation, and returns the appropriate result.
-     *
-     * The [ReturnKind.STRUCTURED] branch delegates parsing, memory, and
-     * observation completion to the structured response coordinator.
-     */
-    private suspend fun finalizeResumedOperation(
-        operation: OperationDefinition,
-        loopResult: ProviderCallResult,
-        messages: MutableList<Message>,
-        correlationId: String,
-        securityContext: ExecutionSecurityContext,
-        conversationId: String?,
-        historySize: Int,
-    ): Any? {
-        when (operation.returnKind) {
-            ReturnKind.STRING -> {
-                // Enforce BEFORE_RESPONSE_RETURN (Fix 3: per-return-kind, not before dispatch)
-                policyHelper.enforce(
-                    policyHelper.buildContext(
-                        enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_RESPONSE_RETURN,
-                        correlationId = correlationId,
-                    ).providerId(loopResult.providerId)
-                        .modelName(loopResult.modelName)
-                        .applySecurityContext(securityContext)
-                        .build()
-                )
-                // Memory persistence + observation + return (once)
-                if (conversationId != null) {
-                    conversationMemoryCoordinator.persistTurn(
-                        PersistConversationTurnRequest(
-                            conversationId,
-                            messages,
-                            historySize,
-                            Message(
-                                role = MessageRole.ASSISTANT,
-                                content = loopResult.response.content,
-                                toolCalls = loopResult.response.toolCalls,
-                            ),
-                        ),
-                    )
-                }
-                loopResult.observation.onCallCompleted(parseSuccess = null)
-                return loopResult.response.content
-            }
-            ReturnKind.UNIT -> {
-                // Enforce BEFORE_RESPONSE_RETURN (Fix 3: per-return-kind, not before dispatch)
-                policyHelper.enforce(
-                    policyHelper.buildContext(
-                        enforcementPoint = dev.tramai.core.policy.EnforcementPoint.BEFORE_RESPONSE_RETURN,
-                        correlationId = correlationId,
-                    ).providerId(loopResult.providerId)
-                        .modelName(loopResult.modelName)
-                        .applySecurityContext(securityContext)
-                        .build()
-                )
-                if (conversationId != null) {
-                    conversationMemoryCoordinator.persistTurn(
-                        PersistConversationTurnRequest(
-                            conversationId,
-                            messages,
-                            historySize,
-                            Message(
-                                role = MessageRole.ASSISTANT,
-                                content = loopResult.response.content,
-                                toolCalls = loopResult.response.toolCalls,
-                            ),
-                        ),
-                    )
-                }
-                loopResult.observation.onCallCompleted(parseSuccess = null)
-                loopResult.response.content // consume it
-                return Unit
-            }
-            ReturnKind.STRUCTURED -> {
-                return structuredResponseCoordinator.finalizeResumed(
-                    ResumedStructuredResponseRequest(
-                        operation = operation,
-                        loopResult = loopResult,
-                        messages = messages,
-                        correlationId = correlationId,
-                        securityContext = securityContext,
-                        conversationId = conversationId,
-                        historySize = historySize,
-                    ),
-                )
-            }
-            ReturnKind.STREAMING -> throw ConfigurationException("Streaming approval resume not supported")
-        }
-    }
-
-    /**
-     * Continues the provider loop after a suspended tool has been executed on resume.
-     *
-     * 1. Enforces BEFORE_TOOL_RESULT_REINJECTION policy
-     * 2. Formats and sanitizes the tool result message
-     * 3. Appends the tool message to the messages list
-     * 4. Processes any remaining unprocessed tool calls from the same batch
-     * 5. Continues the provider loop via [executeWithTools]
-     */
-    private suspend fun continueAfterToolResult(request: ContinueAfterToolResultRequest): ProviderCallResult {
-        val operation = request.operation
-        val messages = request.messages
-        val toolResult = request.toolResult
-        val toolCallId = request.toolCallId
-        val correlationId = request.correlationId
-        val securityContext = request.securityContext
-        val suspendedToolName = request.suspendedToolName
-
-        toolReinjectionCoordinator.reinjectKnownResult(
-            request = ToolCallBatchRequest(
-                operation = operation,
-                messages = messages,
-                toolCalls = emptyList(),
-                correlationId = correlationId,
-                securityContext = securityContext,
-                identity = request.identity,
-                tokenBudgetTracker = request.tokenBudgetTracker,
-                conversationId = request.conversationId,
-                historySize = request.historySize,
-                resumingApproval = request.resumingApproval,
-                parentApprovalId = request.approvalId,
-            ),
-            toolCallId = toolCallId,
-            toolName = suspendedToolName,
-            toolResult = toolResult,
-        )
-        processRemainingResumeToolCalls(request)
-
-        return executeWithTools(
-            ToolLoopContext(
-                operation = operation,
-                messages = messages,
-                tokenBudgetTracker = request.tokenBudgetTracker,
-                correlationId = correlationId,
-                securityContext = securityContext,
-                identity = request.identity,
-                conversationId = request.conversationId,
-                historySize = request.historySize,
-                resumingApproval = request.resumingApproval,
-                parentApprovalId = request.approvalId,
-            ),
-        )
-    }
-
-    private suspend fun processRemainingResumeToolCalls(request: ContinueAfterToolResultRequest) {
-        if (request.toolCallIndex < 0) return
-        val allToolCalls = request.messages
-            .lastOrNull { it.role == MessageRole.ASSISTANT && it.toolCalls != null }
-            ?.toolCalls
-            ?: emptyList()
-        val remainingToolCalls = allToolCalls.drop(request.toolCallIndex + 1)
-        for ((remainingIdx, toolCall) in remainingToolCalls.withIndex()) {
-            appendRemainingResumeToolResult(
-                request = request,
-                toolCall = toolCall,
-                actualIndex = request.toolCallIndex + 1 + remainingIdx,
-            )
-        }
-    }
-
-    private suspend fun appendRemainingResumeToolResult(
-        request: ContinueAfterToolResultRequest,
-        toolCall: ToolCall,
-        actualIndex: Int,
-    ) {
-        try {
-            toolReinjectionCoordinator.processOne(
-                request = ToolCallBatchRequest(
-                    operation = request.operation,
-                    messages = request.messages,
-                    toolCalls = listOf(toolCall),
-                    correlationId = request.correlationId,
-                    securityContext = request.securityContext,
-                    identity = request.identity,
-                    tokenBudgetTracker = request.tokenBudgetTracker,
-                    conversationId = request.conversationId,
-                    historySize = request.historySize,
-                    resumingApproval = request.resumingApproval,
-                    parentApprovalId = request.approvalId,
-                ),
-                toolCall = toolCall,
-                toolCallIndex = actualIndex,
-            )
-        } catch (e: dev.tramai.core.exception.NestedApprovalNotSupportedException) {
-            throw dev.tramai.core.exception.NestedApprovalNotSupportedException(
-                approvalId = request.approvalId,
-                message = "Nested approval not supported in v1: sibling tool ${toolCall.name} requires approval",
-            )
-        } catch (e: ApprovalSuspendedException) {
-            approvalLifecycleAuditEmitter.onUncertainOutcome(
-                approvalId = request.approvalId,
-                workflowRunId = request.identity.workflowRunId,
-                toolName = toolCall.name,
-                reason = "nested-approval-not-supported: sibling tool ${toolCall.name} requires approval",
-            )
-            throw ConfigurationException("Nested approval not supported in v1: sibling tool ${toolCall.name} requires approval")
-        }
-    }
-
-    private data class ContinueAfterToolResultRequest(
-        val operation: OperationDefinition,
-        val messages: MutableList<Message>,
-        val toolResult: ToolResult,
-        val toolCallId: String,
-        val toolCallIndex: Int,
-        val correlationId: String,
-        val securityContext: ExecutionSecurityContext,
-        val identity: EngineExecutionIdentity,
-        val tokenBudgetTracker: TokenBudgetTracker,
-        val suspendedToolName: String = "",
-        val approvalId: String = "",
-        val conversationId: String? = null,
-        val historySize: Int = 0,
-        val resumingApproval: Boolean = false,
-    )
-
-    private fun handleObjectMethod(
-        proxy: Any,
-        method: Method,
-        args: Array<out Any?>,
-    ): Any? = when (method.name) {
-        "toString" -> "TramaiProxy(${serviceDefinition.serviceType.qualifiedName})"
-        "hashCode" -> System.identityHashCode(proxy)
-        "equals" -> proxy === args.firstOrNull()
-        else -> throw UnsupportedOperationException("Unsupported Object method: ${method.name}")
-    }
-
-}
 
 data class OperationDefinition(
     val method: Method,
@@ -1980,7 +950,6 @@ internal fun buildOperationCacheKeyForTesting(
     )
 }
 
-private typealias ProviderCallResult = dev.tramai.engine.provider.ProviderCallResult
 
 
 internal open class ProviderCircuitBreaker(
@@ -2064,10 +1033,6 @@ enum class ReturnKind {
     STREAMING,
 }
 
-private fun PolicyContextBuilder.applySecurityContext(
-    securityContext: ExecutionSecurityContext,
-): PolicyContextBuilder = dataClassification(securityContext.dataClassification)
-    .classificationSource(securityContext.classificationSource)
 
 private fun ExecutionSecurityContext.toCacheSecurityPartition() = CacheSecurityPartition(
     dataClassification = dataClassification,
@@ -2082,12 +1047,6 @@ internal fun sha256Hex(input: String): String {
 }
 
 
-private const val ATTR_RETRY_INDEX = "retry_index"
-private const val ATTR_DELAY_MILLIS = "delay_millis"
-private const val ATTR_DELAY_SOURCE = "delay_source"
-private const val ATTR_LIMIT_TOKENS = "limit_tokens"
-private const val ATTR_OBSERVED_TOKENS = "observed_tokens"
-private const val ATTR_SCOPE = "scope"
 
 /** @see TramaiEngine */
